LPD-85155 extract performance test
