+++ modules/apps/batch-engine/batch-engine-test/src/testIntegration/java/com/liferay/batch/engine/internal/test/BatchEngineExportTaskExecutorTest.java
@@ -27,14 +27,12 @@ import com.liferay.petra.io.unsync.UnsyncBufferedReader;
 import com.liferay.petra.string.CharPool;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.spring.orm.LastSessionRecorderHelper;
 import com.liferay.portal.kernel.spring.orm.LastSessionRecorderHelperUtil;
 import com.liferay.portal.kernel.test.AssertUtils;
 import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.test.TestInfo;
 import com.liferay.portal.kernel.test.rule.DataGuard;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
@@ -101,57 +99,6 @@ public class BatchEngineExportTaskExecutorTest
 		_parameters = HashMapBuilder.<String, Serializable>put(
 			"siteId", TestPropsValues.getGroupId()
 		).build();
-	}
-
-	@Test
-	@TestInfo("LPD-85155")
-	public void testExportBlogPostingsPerformance() throws Exception {
-		Group group = GroupTestUtil.addGroup();
-
-		for (int i = 0; i < 3000; i++) {
-			blogsEntryLocalService.addEntry(
-				user.getUserId(), "headline" + i, "alternativeHeadline" + i,
-				null, "articleBody" + i, new Date(baseDate.getTime()), false,
-				false, null, null, null, null,
-				ServiceContextTestUtil.getServiceContext(
-					TestPropsValues.getCompanyId(), group.getGroupId(),
-					user.getUserId()));
-		}
-
-		batchReadDurations.clear();
-
-		_batchEngineExportTask =
-			_batchEngineExportTaskLocalService.addBatchEngineExportTask(
-				null, user.getCompanyId(), user.getUserId(), null,
-				BlogPosting.class.getName(), "JSON",
-				BatchEngineTaskExecuteStatus.INITIAL.name(),
-				Arrays.asList("headline", "id"),
-				HashMapBuilder.<String, Serializable>put(
-					"siteId", group.getGroupId()
-				).build(),
-				null);
-
-		_batchEngineExportTaskExecutor.execute(_batchEngineExportTask);
-
-		int batchCount = batchReadDurations.size();
-
-		long first5Average = 0;
-
-		for (int i = 0; i < 5; i++) {
-			first5Average += batchReadDurations.get(i);
-		}
-
-		first5Average = first5Average / 5;
-
-		long last5Average = 0;
-
-		for (int i = batchCount - 5; i < batchCount; i++) {
-			last5Average += batchReadDurations.get(i);
-		}
-
-		last5Average = last5Average / 5;
-
-		Assert.assertTrue(last5Average <= (first5Average * 2));
 	}
 
 	@Test
